+++ CHANGELOG.md
@@ -22,10 +22,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security 
 -->
 
+## [0.1.0a10] - 2026-08-13
+
+### Added
+- many more example tests
+
+### Fixed
+- **kwargs format
+
+
 ## [0.1.0a9] - 2026-08-11
 
 ### Fixed
 - bad indent when formatting with `|`
+- **kwargs format
 
 
 ## [0.1.0a8] - 2026-08-10